# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/clone:
    post:
      operationId: cloneVoice
      tags: [voice]
      summary: Clone user voice
      description: Accepts an audio file and creates a cloned voice via ElevenLabs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
              required:
                - audio
      responses:
        "200":
          description: Voice cloned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloneVoiceResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CloneVoiceResponse:
      type: object
      properties:
        voiceId:
          type: string
        message:
          type: string
      required:
        - voiceId
        - message
...